+++ .gitignore
@@ -2,5 +2,4 @@ Gemfile.lock
 *.gem
 .bundle
 .rvmrc
-test/config/database.yml
 test/test.log
