+++ .travis.yml
@@ -5,11 +5,6 @@ rvm:
   - 2.1.8
   - 2.2.4
   - 2.3.0
-before_script:
-  - cp test/config/database.yml.example test/config/database.yml
-  - mysql -e 'create database global_uid_test;'
-  - mysql -e 'create database global_uid_test_id_server_1;'
-  - mysql -e 'create database global_uid_test_id_server_2;'
 gemfile:
   - gemfiles/rails3.2.gemfile
   - gemfiles/rails4.0.gemfile
