+++ gemfiles/rails3.2.gemfile
@@ -4,3 +4,4 @@ gemspec :path=>"../"
 
 gem "activerecord", "~> 3.2.18"
 gem "minitest", "~> 4.7.5"
+gem "mysql2", "~> 0.3.20"
