+++ gemfiles/rails4.0.gemfile
@@ -3,3 +3,5 @@ source "https://rubygems.org"
 gemspec :path=>"../"
 
 gem "activerecord", "~> 4.0.6"
+gem "minitest", "~> 4.7.5"
+gem "mysql2", "~> 0.3.20"
