+++ gemfiles/rails4.0.gemfile.lock
@@ -8,16 +8,16 @@ PATH
 GEM
   remote: https://rubygems.org/
   specs:
-    activemodel (4.0.6)
-      activesupport (= 4.0.6)
+    activemodel (4.0.13)
+      activesupport (= 4.0.13)
       builder (~> 3.1.0)
-    activerecord (4.0.6)
-      activemodel (= 4.0.6)
+    activerecord (4.0.13)
+      activemodel (= 4.0.13)
       activerecord-deprecated_finders (~> 1.0.2)
-      activesupport (= 4.0.6)
+      activesupport (= 4.0.13)
       arel (~> 4.0.0)
-    activerecord-deprecated_finders (1.0.3)
-    activesupport (4.0.6)
+    activerecord-deprecated_finders (1.0.4)
+    activesupport (4.0.13)
       i18n (~> 0.6, >= 0.6.9)
       minitest (~> 4.2)
       multi_json (~> 1.3)
@@ -25,20 +25,23 @@ GEM
       tzinfo (~> 0.3.37)
     arel (4.0.2)
     builder (3.1.4)
-    bump (0.5.0)
-    i18n (0.6.9)
+    bump (0.5.3)
+    i18n (0.7.0)
     metaclass (0.0.4)
     minitest (4.7.5)
     minitest-rg (1.1.1)
       minitest (< 5.0)
     mocha (1.1.0)
       metaclass (~> 0.0.1)
-    multi_json (1.10.1)
-    mysql2 (0.3.16)
-    rake (10.4.2)
-    thread_safe (0.3.4)
-    tzinfo (0.3.39)
-    wwtd (0.5.3)
+    multi_json (1.11.2)
+    mysql2 (0.3.20)
+    phenix (0.2.0)
+      activerecord (>= 3.2, < 5.1)
+      bundler
+    rake (10.5.0)
+    thread_safe (0.3.5)
+    tzinfo (0.3.46)
+    wwtd (1.3.0)
 
 PLATFORMS
   ruby
@@ -48,10 +51,11 @@ DEPENDENCIES
   bump
   bundler
   global_uid!
-  minitest
+  minitest (~> 4.7.5)
   minitest-rg
   mocha
-  mysql2
+  mysql2 (~> 0.3.20)
+  phenix
   rake
   wwtd (>= 0.5.3)
