+++ gemfiles/rails4.1.gemfile
@@ -3,3 +3,4 @@ source "https://rubygems.org"
 gemspec :path=>"../"
 
 gem "activerecord", "~> 4.1.6"
+gem "mysql2", "~> 0.3.20"
