--- test/config/database.yml.example
@@ -1,20 +1,0 @@
-test:
-  adapter: mysql2
-  encoding: utf8
-  database: global_uid_test
-  username: root
-  password:
-
-test_id_server_1:
-  adapter: mysql2
-  encoding: utf8
-  database: global_uid_test_id_server_1
-  username: root
-  password:
-
-test_id_server_2:
-  adapter: mysql2
-  encoding: utf8
-  database: global_uid_test_id_server_2
-  username: root
-  password:
