+++ test/test_helper.rb
@@ -4,6 +4,7 @@ require 'minitest/autorun'
 require 'minitest/rg'
 require 'mocha/setup'
 require 'global_uid'
+require 'phenix'
 
 GlobalUid::Base.global_uid_options = {
   :disabled   => false,
@@ -14,8 +15,12 @@ GlobalUid::Base.global_uid_options = {
 }
 GlobalUid::Base.extend(GlobalUid::ServerVariables)
 
-yaml = YAML.load(IO.read(File.dirname(__FILE__) + "/config/database.yml"))
-ActiveRecord::Base.configurations = yaml
+Phenix.configure do |config|
+  config.database_config_path = File.join(File.dirname(__FILE__), "config/database.yml")
+end
+
+Phenix.rise!(with_schema: false)
 ActiveRecord::Base.establish_connection(:test)
 ActiveRecord::Base.logger = Logger.new(File.dirname(__FILE__) + "/test.log")
 ActiveSupport.test_order = :sorted if ActiveSupport.respond_to?(:test_order=)
+ActiveRecord::Migration.verbose = false
